settings.py: Disable db_populate on prod
David had put in a check that would keep us from running this in
production (which would nuke all data!) but we had left it broken so we
could simulate prod.  Now that there really is a prod, turn this off for
good.

If we ever want to run in production, we can manually edit settings.py
to override.